requirements.txt
Django==4.2.8
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
psycopg2-binary==2.9.9
celery==5.3.4
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1
WeasyPrint==60.1
Pillow==10.1.0
language-tool-python==2.7.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-magic==0.4.27
bleach==6.1.0
daphne==4.0.0
html2text==2024.2.26
Werkzeug==3.0.1
